+++ frontend/src/App.tsx
@@ -6,20 +6,23 @@ import { Router } from "./router/Router";
 import "./App.css";
 import { HeaderLayout } from "./components/templates/HeaderLayout";
 import { LoginUserProvider } from "./providers/LoginUserProvider";
+import { PostProvider } from "./providers/PostProvider";
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 function App() {
   return (
     <LoginUserProvider>
-      <ChakraProvider theme={theme}>
-        <BrowserRouter>
-          <HeaderLayout>
-            <Box m={16}>
-              <Router />
-            </Box>
-          </HeaderLayout>
-        </BrowserRouter>
-      </ChakraProvider>
+      <PostProvider>
+        <ChakraProvider theme={theme}>
+          <BrowserRouter>
+            <HeaderLayout>
+              <Box m={16}>
+                <Router />
+              </Box>
+            </HeaderLayout>
+          </BrowserRouter>
+        </ChakraProvider>
+      </PostProvider>
     </LoginUserProvider>
   );
 }
